+++ src/components/Layout/components/Header/Header.module.scss
@@ -10,6 +10,7 @@
     align-items: center;
     justify-content: space-between;
     background: transparent;
+    background: linear-gradient(245.59deg, #00a671 0%, #01704d 28.53%, #014f36 75.52%);;
     transition: all .2s ease-in;
 
     &.active {
