+++ src/components/Layout/components/Header/Header2.module.scss
@@ -1,0 +1,116 @@
+
+.wrapper {
+    width: 100%;
+    padding: 24px 0 16px 0;
+    font-family: var(--primary-font);
+    position: fixed;
+    display: flex;
+    z-index: 10;
+    align-items: center;
+    justify-content: center;
+    // background-color: white;
+    // background: linear-gradient(245.59deg, #00a671 0%, #01704d 28.53%, #014f36 75.52%);
+    transition: padding-top .3s ,background-color 0ms;
+    &.active {
+        background-color: white;
+        padding-top: 8.2px;
+        padding-bottom: 8px;
+        box-shadow: 0 .375rem .875rem #0000001a;
+    }
+
+}
+
+.inner {
+    background-color: white;
+    display: flex;
+    align-items: center;
+    width: 75%;
+    justify-content: space-between;
+    padding: 0 16px;
+    height: 6.2rem;
+    border-radius: 50px;
+}
+
+
+
+.button-search {
+    color: var(--text-color);
+    padding: 5px 15px;
+    border: 1px solid transparent;
+    line-height: 1.4;
+    font-size: 1.4rem;
+    border-radius: 15px;
+    transition: all .55s ease;
+    cursor: pointer;
+
+    &:hover {
+        background: #4d9559;
+        transform: translateX(6px);
+    }
+
+    .icon-search {
+        margin-top: 5px;
+        font-size: 2.5rem;
+    }
+}
+
+
+.logo {
+
+
+}
+
+.group {
+    display: flex;
+    align-items: center;
+
+}
+
+.icon-logo {
+    font-size: 30px;
+    font-weight: 800;
+    color: #34ca96;
+    letter-spacing: 0.4rem;
+    text-transform: uppercase;
+    transition: all 0.3s ease-in-out;
+
+
+    &:hover {
+        opacity: 0.8;
+    }
+}
+
+.nav-list {
+    display: flex;
+    list-style: none;
+    gap: 20px;
+    margin-bottom: 0px;
+    padding-left: 0;
+    margin-left: 20px;
+    margin-right: 20px;
+}
+
+.nav-list li a {
+    color: var(--text-color);
+    font-weight: 600;
+    position: relative;
+
+    &::before {
+        position: absolute;
+        bottom: -2px;
+        content: "";
+        width: 0;
+        background-color: #4d9559;
+        height: 2px;
+        right: 0;
+        transition: all .3s ease-in-out;
+    }
+}
+
+.nav-list li:hover a::before {
+    width: 100%;
+}
+
+
+
+
