+++ src/pages/user/Explore/Explore.module.scss
@@ -1,0 +1,3 @@
+.wrapper {
+    height: 10000px;
+}
