+++ src/pages/user/Explore/index.js
@@ -1,0 +1,15 @@
+import classNames from "classnames/bind";
+
+import styles from './Explore.module.scss'
+
+const cx = classNames.bind(styles);
+
+function Explore() {
+    return (
+        <div className={cx('wrapper')}>
+       
+        </div>
+    );
+}
+
+export default Explore;
