+++ src/routes/index.js
@@ -3,12 +3,14 @@ import Following from '~/pages/Following'
 import Upload from '~/pages/Upload'
 import { HeaderOnlyLayout } from '~/components/Layout'
 import { HomeLayout } from '~/components/Layout'
+import Explore from '~/pages/user/Explore'
 
 // Public routes
 const publicRoutes = [
     { path: '/', component: Home, layout: HomeLayout},
     { path: '/following', component: Following},
-    { path: '/upload', component: Upload}
+    { path: '/upload', component: Upload},
+    { path: '/explore/:type', component : Explore, layout: HeaderOnlyLayout }
 ]
 
 const privateRoutes = [
